Split server into lib and thin CLI
